requirements.txt
Flask
opencv-python-headless
numpy
Pillow
tensorflow
matplotlib